+++ components/shared_code/src/shared_data_model/sources/__init__.py
@@ -10,7 +10,6 @@ from .cargo_audit import CARGO_AUDIT
 from .cloc import CLOC
 from .cobertura import COBERTURA, COBERTURA_JENKINS_PLUGIN
 from .composer import COMPOSER
-from .cxsast import CXSAST
 from .dependency_track import DEPENDENCY_TRACK
 from .gatling import GATLING
 from .generic_json import GENERIC_JSON
@@ -58,7 +57,6 @@ SOURCES = {
     "cobertura": COBERTURA,
     "cobertura_jenkins_plugin": COBERTURA_JENKINS_PLUGIN,
     "composer": COMPOSER,
-    "cxsast": CXSAST,
     "dependency_track": DEPENDENCY_TRACK,
     "gatling": GATLING,
     "generic_json": GENERIC_JSON,